refactor(memory qrm plugin): split advisor handler into linux and non-linux versions
Move the cgroup-v1/v2 dependent handleAdvisorMemoryHot implementation
to platform-specific files to support cross-compilation on non-linux systems
