+++ pkg/agent/qrm-plugins/memory/dynamicpolicy/policy_advisor_handler.go
@@ -26,7 +26,6 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/opencontainers/runc/libcontainer/cgroups"
 	"github.com/samber/lo"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
@@ -391,43 +390,6 @@ func (p *DynamicPolicy) handleAdvisorResp(advisorResp *advisorsvc.ListAndWatchRe
 	if err := p.state.StoreState(); err != nil {
 		general.ErrorS(err, "store state failed")
 	}
-	return nil
-}
-
-func (p *DynamicPolicy) handleAdvisorMemoryHigh(
-	_ *config.Configuration,
-	_ interface{},
-	_ *dynamicconfig.DynamicAgentConfiguration,
-	emitter metrics.MetricEmitter,
-	metaServer *metaserver.MetaServer,
-	entryName, subEntryName string,
-	calculationInfo *advisorsvc.CalculationInfo, podResourceEntries state.PodResourceEntries,
-) error {
-	memoryHighStr := calculationInfo.CalculationResult.Values[string(memoryadvisor.ControlKnobKeyMemoryHigh)]
-	memoryHigh, err := strconv.ParseInt(memoryHighStr, 10, 64)
-	if err != nil {
-		return fmt.Errorf("parse %s: %s failed with error: %v", memoryadvisor.ControlKnobKeyMemoryHigh, memoryHighStr, err)
-	}
-
-	if !cgroups.IsCgroup2UnifiedMode() {
-		general.Infof("memory.high is not supported in cgroupv1 mode")
-		return nil
-	}
-
-	if calculationInfo.CgroupPath != "" {
-		if err = cgroupmgr.ApplyMemoryWithRelativePath(calculationInfo.CgroupPath, &common.MemoryData{
-			HighInBytes: memoryHigh,
-		}); err != nil {
-			return fmt.Errorf("apply memory.high failed with error: %v", err)
-		}
-
-		_ = emitter.StoreInt64(util.MetricNameMemoryHandleAdvisorMemoryHigh, memoryHigh,
-			metrics.MetricTypeNameRaw, metrics.ConvertMapToTags(map[string]string{
-				"cgroupPath": calculationInfo.CgroupPath,
-			})...)
-		return nil
-	}
-
 	return nil
 }
